Announce the movement to a new repository, kamikazi-core.
This kamikazi-deploy repository was useful for most of the Upstart based
ubuntu derived distributions, but as of 15.04, systemd is now the default.

Please visit the new repository at
https://github.com/kamilion/kamikazi-core
for more information, updated versions, and removal of legacy code that
kamikazi-deploy based ISOs require to be in place for backwards compatibility
with the 0.4.0, 0.5.0, 0.6.x, and 0.7.x releases.

As such, it's time to retire this repository, and the ISO images that relied
upon it's existance for continuing updates, and boot-time configuration.

It's been a great run for the last three years, thanks to all who have contributed.