Populate the repository with a default license.
Obviously, if it's required to be GPLv2 or GPLv3 or BSD or whatever
to be compliant with whatever upstream package it's part of?
yeah, use that license. Anybody else's code I reference is theirs.
All credit due to authors of code. All I've done is mess with some
config files and write a few procedural scripts. Your milage may vary.