Something in bash changed from 16.04 to 18.04, now bootscripts fail with error
break: only meaningful in a `for', `while', or `until' loop

So then move the timeout, which was a function, into the while loop itself.
So much for DRY (Don't Repeat Yourself), but I guess that was never bash's strong point anyway.
Hopefully this should work around the issue and restore the previous script behavior.